Refactor Data into a class
This helps to encapsulate some logic inside the class, such as the
static check "is this type T a valid Data type ?", or the translation
from a Data type to a pretty type name suited to display in script
errors.

This commit also refactor the Tree class, by using iterator traversal
and removing not used methods.